backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.1
google-generativeai>=0.3.1
firebase-admin>=6.2.0
